Discover archived log files based on "days back"
Rather than using a fixed set of archived log files, we discover them
by searching for files with a given prefix and of a given (maximum)
age.  This is more robust against different styles of archiving and
compressing log files.

It also avoids errors on systems that haven't accumulated the expected
number of archived log files yet, and makes the tool more useful for
freshly installed systems (with DOA disks).

Addresses #7