Use "sas2ircu" command to physically locate disk
The sas2ircu command can be used to enumerate SAS controllers and the
disks attached to them, and to obtain information such as the physical
enclosure and the slot within an enclosure where a disk is located.  If
the script manages to obtain that information for the disk in question,
it will now output the sas2ircu commands to turn the locator LED on and
off, which is very handy for disk replacement.

Addresses #1