Initial support for LSI/Broadcom SAS 3008 adapters
We simply try to use "sas3ircu" before "sas2ircu".

The current state is ugly because the sas2ircu code is basically just
duplicated.  We should at least "DRY" it by factoring out the common
parts.

Also, it would be nice to make an educated guess for which version of
the tool is available and likely to work.